Rename *object.* into *objecttest.*
